requirements.txt
comtypes==1.1.11
cvzone==1.5.6
cx_Freeze==6.11.1
face_recognition==1.3.0
googletrans==4.0.0rc1
numpy==1.22.1
openai==0.22.1
opencv_python==4.5.5.62
PyAutoGUI==0.9.53
pycaw==20220416
pygame==2.1.2
pyserial==3.5
requests==2.28.1
Unidecode==1.3.4
